Removed old passcode vc, cleared
